Calculate gains aka correction factors
This puts in place the basic machinery of a queue of heaps, serviced
by methods that calculate the necessary products. The first method
only focuses on the real-time / fast "correction factors" aka HH and VV
gains. It only works when the noise diode is firing, and extracts
ON and OFF spectra with proper scaling for the accumulation length.
It estimates the gain as (ON - OFF) / MODEL, where the model is
currently fixed at 20 K. Take the median gain across the channels
to mitigate RFI. Also hardcode the digitiser metadata for now.