Replace mean with median on gain estimate
This reduces sensitivity to outlier channels that can fluctuate a lot,
even between ON and OFF segments. Keep the Tsys based on the mean,
though, as this will include all sources of power. Do a mean across
time but sum across frequency to get to total power.

Also make N_CHANS a constant for now. It cleans up the code a bit.
We can always revert once there are more options for the spectrometer.