Add post-processing correction products
According to requirement R.SP.OP.19, the SP "shall calculate and record
as metadata, based on the noise diode spectra of R.SP.OP.17,
time-averaged and low-order polynomial models of the noise diode power
and XY phase spectra per receptor." This commit adds this functionality.

The raw HH, VV, ReVH and ImVH data products are collected per L0 dump
and averaged in time for the duration of the dump (and divided by the
number of accumulations involved in each). The low-order polynomial
model is interpreted to be a B-spline, which allows efficient fitting
of multiple bumps in the data without numerical issues. The spline
order is fixed to three and the knot coordinates are also precomputed
to save storage space at the cost of slightly lower quality. Each
spectrum of 128 channels is thereby compressed to 21 floats, while
reducing the floating-point precision as well.

Use the empirical standard deviation observed across time within each
frequency channel to weight the spline fits. This automatically
suppresses RFI with no further need for flagging. In the case of the
VH phase we get a rough estimate of its standard deviation by comparing
the VH correlation deviation with the mean VH value, sort of like
calculating radians.